+++ R/reshape.R
@@ -1,16 +1,20 @@
 .reshape_const_MESSAGE_UNCLEANED <- "Logger {serial_number} isn't cleaned. I can't detect the last time_to."
+.reshape_const_MESSAGE_WIDE_UNCLEANED <- "Only one logger per myClim object can be reshaped in uncleaned data."
 
 #' Export values to wide table
 #'
 #' This function converts myClim object to the R data.frame with values of sensor in wide format.
 #'
 #' @details First column of the output data.frame is datetime followed by the
-#' columns for every sensor. Name of the column is in format:
-#' * localityid_loggerid_serialnumber_sensorname for Raw-format and `show_logger_name=FALSE`
-#' * localityid_loggername_sensorname for Raw-format and `show_logger_name=TRUE`
-#' * localityid_sensorname for Agg-format
+#' columns for every sensor. When reshaping uncleaned data in Raw-format, only one logger
+#' per myClim object is allowed to avoid potential confusion in case, there are 
+#' duplicated values, allowed in uncleaned Raw myClim format.
+#' Name of the column is in format:
+#' * `localityid_loggerid_serialnumber_sensorname` for Raw-format and `show_logger_name=FALSE`
+#' * `localityid_loggername_sensorname`for Raw-format and `show_logger_name=TRUE`
+#' * `localityid_sensorname` for Agg-format
 #' 
-#' The less complex wide table is returned when exporting single sensor ascross localities. 
+#' The less complex wide table is returned when exporting single sensor across localities. 
 #'
 #' @template param_myClim_object
 #' @template param_localities
@@ -22,22 +26,34 @@
 #' * locality1_sensor1
 #' * ...
 #' * ...
-#' * localityn_sensorn
+#' * localityN_sensorN
 #' 
 #' @export
 #' @examples
-#' example_tms_wideformat <- mc_reshape_wide(mc_data_example_raw, c("A6W79", "A2E32"),
+#' example_tms_wideformat <- mc_reshape_wide(mc_data_example_clean, c("A6W79", "A2E32"),
 #'                                           c("TMS_T1", "TMS_T2"))
 mc_reshape_wide <- function(data, localities=NULL, sensors=NULL, use_utc=TRUE, show_logger_name=FALSE) {
-    data <- mc_filter(data, localities, sensors)
+    data <- mc_filter(data, localities=localities, sensors=sensors)
     if(.common_is_agg_format(data)) {
         use_utc <- .common_check_agg_use_utc(use_utc, data$metadata@period)
     }
-    datetimes <- .reshape_get_all_datetimes(data, use_utc)
-    tables <- c(tibble::tibble(datetimes), .reshape_get_sensor_tables(data, use_utc, show_logger_name))
-    tables[[1]] <- tibble::as_tibble(tables[[1]])
-    colnames(tables[[1]]) <- "datetime"
-    as.data.frame(purrr::reduce(tables, function(.x, .y) dplyr::left_join(.x, .y, by="datetime")))
+    sensor_tables <- .reshape_get_sensor_tables(data, use_utc, show_logger_name)
+    if(length(sensor_tables) == 0) {
+        return(data.frame())
+    }
+    if(!.prep_is_datetime_step_processed_in_object(data) && length(sensor_tables) > 1) {
+        stop(.reshape_const_MESSAGE_WIDE_UNCLEANED)
+    }
+    if(.prep_is_datetime_step_processed_in_object(data)) {
+        datetimes <- .reshape_get_all_datetimes(data, use_utc)
+        tables <- c(tibble::tibble(datetimes), sensor_tables)
+        tables[[1]] <- tibble::as_tibble(tables[[1]])
+        colnames(tables[[1]]) <- "datetime"
+        result <- as.data.frame(purrr::reduce(tables, function(.x, .y) dplyr::left_join(.x, .y, by="datetime")))
+    } else {
+        result <- as.data.frame(sensor_tables[[1]])
+    }
+    return(result)
 }
 
 .reshape_get_all_datetimes <- function(data, use_utc) {
@@ -123,7 +139,7 @@ mc_reshape_wide <- function(data, localities=NULL, sensors=NULL, use_utc=TRUE, s
 #' @examples
 #' head(mc_reshape_long(mc_data_example_clean, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2")), 10)
 mc_reshape_long <- function(data, localities=NULL, sensors=NULL, use_utc=TRUE) {
-    data <- mc_filter(data, localities, sensors)
+    data <- mc_filter(data, localities=localities, sensors=sensors)
     is_raw_format <- .common_is_raw_format(data)
     period <- NULL
     if (!is_raw_format && !(data$metadata@period %in% .agg_const_INTERVAL_PERIODS)) {