+++ tests/testthat/test-read.R
@@ -74,6 +74,7 @@ test_that("mc_read_files TOMST directory", {
     expect_equal(length(data$localities[[1]]$loggers), 1)
     expect_equal(data$localities$`92192250`$loggers[["Dendro_1"]]$metadata@type, .model_const_LOGGER_TOMST_DENDROMETER)
     expect_equal(length(data$localities$`92192250`$loggers[["Dendro_1"]]$sensors), 2)
+    expect_equal(data$localities$`94184102`$loggers[["TMS_2"]]$metadata@raw_index[[1]], 8316)
 })
 
 test_that("mc_read_data TOMST 2024 format changes", {
@@ -309,4 +310,15 @@ test_that("mc_read_files HOBO logger_format", {
     test_raw_data_format(data)
     expect_equal(names(data$localities$A$loggers[["HOBO_U23-001A_1"]]$sensors), c("HOBO_T", "HOBO_RH"))
     expect_equal(names(data$localities$A$loggers[["HOBO_U23-004_1"]]$sensors), c("HOBO_T", "HOBO_extT"))
+})
+
+test_that("mc_read_files joined TOMST logger type", {
+    data <- mc_read_files("../data/joined_TOMST/problems/202310_94218703.csv", "TOMST_join", silent=TRUE)
+    expect_equal(data$localities$`202310_94218703`$loggers[["Thermo_1"]]$metadata@type, "Thermo")
+    expect_equal(names(data$localities$`202310_94218703`$loggers[["Thermo_1"]]$sensors), "Thermo_T")
+    data <- mc_read_files("../data/joined_TOMST/problems/202310_94218703.csv", "TOMST_join",
+                          logger_type="TMS_L45", silent=TRUE)
+    expect_equal(data$localities$`202310_94218703`$loggers[["TMS_L45_1"]]$metadata@type, "TMS_L45")
+    expect_equal(names(data$localities$`202310_94218703`$loggers[["TMS_L45_1"]]$sensors),
+                 c("TMS_T1_L45", "TMS_T2_L45", "TMS_T3_L45", "TMS_moist_L45", "VWC"))
 })