+++ tests/testthat/test-reshape.R
@@ -1,10 +1,13 @@
 test_that("wideformat-filter", {
     data <- mc_read_data("../data/TOMST/files_table.csv", clean=FALSE)
-    table <- mc_reshape_wide(data, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2"))
+    expect_error(table <- mc_reshape_wide(data, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2")))
+    table <- mc_reshape_wide(data, "A6W79", c("TMS_T1", "TMS_T2"))
+    expect_true("A6W79_1_94184102_TMS_T1" %in% colnames(table))
+    cleaned_data <- mc_prep_clean(data, silent=T)
+    table <- mc_reshape_wide(cleaned_data, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2"))
     expect_true("A6W79_1_94184102_TMS_T1" %in% colnames(table))
     expect_equal(ncol(table), 5)
     expect_equal(nrow(table), 100)
-    cleaned_data <- mc_prep_clean(data, silent=T)
     agg_data <- mc_agg(cleaned_data)
     table <- mc_reshape_wide(agg_data, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2"))
     expect_true("A6W79_TMS_T1" %in% colnames(table))
@@ -13,7 +16,7 @@ test_that("wideformat-filter", {
 })
 
 test_that("wideformat-all", {
-    data <- mc_read_data("../data/TOMST/files_table.csv", clean=FALSE)
+    data <- mc_read_data("../data/TOMST/files_table.csv", clean=TRUE, silent=TRUE)
     table <- mc_reshape_wide(data)
     expect_equal(ncol(table), 10)
     expect_equal(nrow(table), 111)
@@ -21,20 +24,19 @@ test_that("wideformat-all", {
 
 test_that("wideformat-all local time", {
     data <- mc_read_data("../data/TOMST/files_table.csv",
-                         localities_table = "../data/TOMST/localities_table.csv", clean=FALSE)
+                         localities_table = "../data/TOMST/localities_table.csv", clean=TRUE, silent=TRUE)
     table_utc <- mc_reshape_wide(data, use_utc=TRUE)
     table <- mc_reshape_wide(data, use_utc=FALSE)
     expect_equal(table_utc$A1E05_91184101_Thermo_T[table_utc$datetime == lubridate::ymd_h("2020-10-28 9")],
                  table$A1E05_91184101_Thermo_T[table$datetime == lubridate::ymd_h("2020-10-28 10")])
     table$datetime <- table$datetime - (60 * 60)
     expect_equal(table_utc, table)
-    cleaned_data <- mc_prep_clean(data, silent=T)
-    agg_data <- mc_agg(cleaned_data)
+    agg_data <- mc_agg(data)
     table_utc <- mc_reshape_wide(agg_data, use_utc=TRUE)
     table <- mc_reshape_wide(agg_data, use_utc=FALSE)
     table$datetime <- table$datetime - (60 * 60)
     expect_equal(table_utc, table)
-    agg_data <- mc_agg(cleaned_data, fun="coverage", period="day")
+    agg_data <- mc_agg(data, fun="coverage", period="day")
     expect_warning(table <- mc_reshape_wide(agg_data, use_utc=FALSE))
 })