+++ tests/testthat/test-states.R
@@ -418,20 +418,32 @@ test_that("mc_states_join", {
     data <- mc_read_files("../data/join_tolerance", "TOMST", silent=TRUE)
     states_data <- mc_states_join(data)
     states_table <- mc_info_states(states_data) %>% dplyr::filter(tag == "join_conflict")
-    expect_equal(nrow(states_table), 8)
+    expect_equal(nrow(states_table), 10)
     expect_true("TMS_1(94184101)" %in% states_table$value)
     expect_true("TMS_2(94184101)" %in% states_table$value)
+    expect_true("TMS_3(94184101)" %in% states_table$value)
+    filtered_states_table <- dplyr::filter(states_table, .data$logger_name == "TMS_2" & .data$sensor_name == "TMS_T1")
+    expect_equal(nrow(filtered_states_table), 2)
     tolerance <- list(T_C=0.5)
     states_data <- mc_states_join(data, tolerance=tolerance)
     states_table <- mc_info_states(states_data) %>% dplyr::filter(tag == "join_conflict")
     expect_equal(nrow(states_table), 2)
 })
 
-
 test_that("mc_states_join no states", {
     data <- mc_read_files("../data/join_tolerance", "TOMST", silent=TRUE)
     data <- mc_calc_vwc(data)
     states_data <- mc_states_join(data)
     states_table <- mc_info_states(states_data) |> dplyr::filter(tag == "join_conflict")
-    expect_equal(nrow(states_table), 10)
+    expect_equal(nrow(states_table), 14)
 })
+
+test_that("mc_states_join suffix", {
+    data <- mc_read_files("../data/join_tolerance", "TOMST", silent=TRUE)
+    states_data <- mc_states_join(data, older_newer_suffix = TRUE)
+    older_states_table <- mc_info_states(states_data) %>% dplyr::filter(tag == "join_conflict_older")
+    newer_states_table <- mc_info_states(states_data) %>% dplyr::filter(tag == "join_conflict_newer")
+    expect_equal(nrow(older_states_table), 5)
+    expect_equal(nrow(newer_states_table), 5)
+})
+